internal(release): render release pr body
